This check-in of 'runwp2driz.py' includes the following changes:     - minor spelling error in 'help()' fixed     - call to SkyField corrected     - support for *0m.fits name for MEF files in pipeline added     - added CTE keyword computation task to the script           (Ticket 179 and OPR 58998)
This check-in also updated 'wfpc2cte.py' in 'wfpc2tools' package to
Version 1.1 (29-Feb-2008).  It will now correctly convert the GEIS
input to MEF so that it updates only the MEF file, and it will now correctly flush the updated header from memory to the MEF file.

The updated script and  was successfully tested on 'basil' (Solaris 10)
under IRAFDEV [using refactored PyDrizzle] with u40x0102m_c0f.fits
as input. Testing verified that the product came out with the right
pixel scale and the trailer file was updated properly.

WJH




git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/wfpc2tools@1183 fe389314-cf27-0410-b35b-8c050e845b92